# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "homework API",
        "contact": {},
        "license": {
            "name": "ymd"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "アプリケーションのログイン",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ログインAPI",
                "parameters": [
                    {
                        "description": "サインアップ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.LonginResponse"
                        }
                    }
                }
            }
        },
        "/signin": {
            "post": {
                "description": "ユーザー新規登録",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ユーザー新規登録API",
                "parameters": [
                    {
                        "description": "サインアップ",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "user.LonginResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "user.UserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    }
}